requirements.txt
matplotlib
numpy
opencv-python
pandas
Pillow
plotly
PyYAML
requests
streamlit
tensorflow
tensorflow-intel
torch
torchvision
transformers
ultralytics
huggingface-hub
datetime

